Dette blacksmith inventoy (#13)
* c

c

* dd

c

* Fix Player controls disabled when menu opened

* c

c

* debut hover

c

* partie0.3 hover item

c

* c

c

* c

c

* c

c

* c

c

* ref inventory slot index

* 1er test fusion inventaire hover

* c

c

* c

c

* fix drag and drop blacksmith temp

* c

c

* hovver fonctionnel part 1

* cc

c

* Multiple changes

- Added a way to Freeze/Unfreeze controllers, toggling their update
- Added a UIComponent class
- Changed UIMenu to use UIComponent
- Rewrote NavBar and NavTab

* Fixed + Cleaned player

* .

* Hover

* .

* IHoverable

* Drag & Drop

* Documentation

* Update

* UI changes

* correction reference

* c

c

* Changes

- Reverted the changes of the ControllerManager
- Added a proper API for the UIManager
- cried

* drag and drop gem

* addsloot

* Moved IElementable

* drag and drop gem inventory

* c

c

* c

c

* fix drag and drop with add item to index

* .

* debut stat upgrade

* upgrade 0.2

* Multiple changes

- Moved all files from Assets/Modules/Blacksmyth to Assets/Modules/Blacksmith
- Improved the display of the upgrade stats
- Moved the upgrade stats display to it's own class

* Fixed the UI for ShowGemShape

* Changes

- Added a visual that tells the player which case was active inside the blacksmith
- Added an interface to any class to use InputMaster
- Modified the documentation of the action interfaces to include all objects
- Modified Controller to use the new interface IActionable
- Made UIComponent an IActionable by default

* c

c

* ROTATE FOR THGE GLORRYYYYYYYYYYYYYYYYYYYYYYYY

* c

c

* test

test

* fusion Ui :)

* Hover Tab

- Fixed a bug with HoverManager
- Updated the Input System
- Added a Tab functionality to GemHover
-  Added new methods for NavBar

* fusion 0.5

* fusion fonctionel

* c

c

* Better Fuse UI

* c

c

* c

c

---------

+++ Assets/Modules/Fishing/Scripts/FishingManager.cs
@@ -35,7 +35,7 @@ namespace FishingModule
             if (isCollected)
             {
               foreach (KeyValuePair<FishSO, uint> fish in this._buoy.GetFishCaught())
-                    Inventory.Instance.AddItem(new FishData(fish.Key, fish.Value));
+                    Inventory.Instance.AddItem(new FishData(fish.Key, (int) fish.Value));
             }
             else
             {
